The library compiles under WASM and out of it.
