+++ linera-ethereum/src/provider.rs
@@ -14,11 +14,37 @@ use alloy_primitives::Bytes;
 use async_trait::async_trait;
 use url::Url;
 use crate::client::EthereumQueries;
+use crate::client::JsonRpcClient;
+use reqwest::Client as Client_json_ser;
 
 pub type HttpProvider = RootProvider<alloy::transports::http::Http<Client>>;
 
 use crate::common::{event_name_from_expanded, parse_log, EthereumEvent, EthereumServiceError};
 use crate::client::get_block_id;
+
+/// The Ethereum endpoint and its provider used for accessing the ethereum node.
+pub struct EthereumClientSimplified {
+    pub url: String,
+}
+
+#[async_trait]
+impl JsonRpcClient for EthereumClientSimplified {
+    type Error = EthereumServiceError;
+    async fn request_inner(&self, payload: Vec<u8>) -> Result<Vec<u8>, Self::Error> {
+        let client = Client_json_ser::new();
+	let res = client.post(self.url.clone()).json_ser(payload).send().await?;
+        let body = res.bytes().await?;
+        Ok(body.as_ref().to_vec())
+    }
+}
+
+impl EthereumClientSimplified {
+    /// Connects to an existing Ethereum node and creates an `EthereumEndpoint`
+    /// if successful.
+    pub fn new(url: String) -> Self {
+        Self { url }
+    }
+}
 
 pub struct EthereumClient<M> {
     pub provider: M,
